// src/generated/models/Crf82_blockedperiodsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf82_blockedperiodsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf82_blockedperiodsstatecode = keyof typeof Crf82_blockedperiodsstatecode;
export const Crf82_blockedperiodsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf82_blockedperiodsstatuscode = keyof typeof Crf82_blockedperiodsstatuscode;

export interface Crf82_blockedperiodsBase {
  crf82_blockedperiodid: string;
  crf82_campuslocation?: string;
  crf82_comments?: string;
  crf82_enddate?: string;
  crf82_examtitle?: string;
  crf82_legacyidentifier?: number;
  crf82_startdate?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf82_blockedperiodsstatecode;
  statuscode?: Crf82_blockedperiodsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf82_blockedperiods extends Crf82_blockedperiodsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}